refactor(ogc): remove chunking dependency hop
